マイグレーション修正
