+++ alembic/versions/2023_1023_1532_init.py
@@ -1,0 +1,54 @@
+"""init
+
+Revision ID: 8c5735c87768
+Revises: 
+Create Date: 2023-10-23 15:32:34.893311
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '8c5735c87768'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(length=255), server_default='', nullable=False),
+    sa.Column('email', sa.String(length=255), nullable=True),
+    sa.Column('password', sa.String(), nullable=True),
+    sa.Column('order_num', sa.Integer(), server_default='0', nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('email')
+    )
+    op.create_table('tasks',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(length=255), server_default='', nullable=False),
+    sa.Column('order_num', sa.Integer(), server_default='0', nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('run_dates',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('task_id', sa.Integer(), nullable=False),
+    sa.Column('run_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.Column('order_num', sa.Integer(), server_default='0', nullable=False),
+    sa.ForeignKeyConstraint(['task_id'], ['tasks.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('run_dates')
+    op.drop_table('tasks')
+    op.drop_table('users')
+    # ### end Alembic commands ###
